[SCHEMA] Add Message.namespace_id column, part 1.
Summary:
This is important both for performance and for assurance of namespace
isolation. In this commit, we merely add the column without requiring that it
be non-null. This is so that we can run the (time-consuming) migration online,
and then alter the column to be non-nullable in a subsequent (faster)
migration.

Associated cleanups:
* Move the contents of Message.__init__, which previously was responsible
  for constructing a Message instance from synced data, to a
  Message.create_from_synced method for greater clarity. Remove unused flags
  argument.
* Remove Message.create_draft_message, which wasn't really doing much.

Test Plan: Run syncs, unit tests.

Reviewers: charles

Differential Revision: https://review.inboxapp.com/D496